Re-organise test_harness binary
Moves manifest and components into separate files.
